+++ src/lib.rs
@@ -236,10 +236,10 @@ impl State {
             }
         }
 
-        // After final boss is defeated, split at the start of the next cutscene
+        // After final boss is defeated, split at the start of the next load
         if game.ready_to_end
             && game_flow_state
-                .map(|x| x.current == GameFlowState::CinematicSequenceState)
+                .map(|x| x.current == GameFlowState::LoadingTransitionState)
                 .unwrap_or(false)
         {
             game.ready_to_end = false;