+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- Moved end condition to start of load (initial loss of control) rather than start of the final cutscene
+
 ## [0.0.5] - 2023-02-05
 
 ## Fixed